requirements.txt
discord.py==1.7.3
validators==0.23.2
